python_projects/023_project/themes/image_editor/requirements.txt
PySimpleGUI>=4.60.0
Pillow>=9.0.0 